Improve iOS test frameworks and testing.
The prior approach I took to iOS testing was to make a full-featured Swift app that allowed blackbox testing, and to build some cursory self-tests into it, but really to rely on the macOS tests for coverage.  This was a bad idea, because the iOS and macOS libraries aren't functionally identical, so it let through some iOS-only bugs.

This time, I have greatly simplified the Swift app to simply be a test harness (sort of like Rust's built-in test harness) that runs a compiled static library of tests.  Then by making this library run all the same tests that are run by Rust's test harness on macOS, I get full coverage of the underlying framework on both platforms.